+++ ChangeLog.md
@@ -2,6 +2,17 @@
 
 All notable changes are documented in this file using the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [8.2.1] - 2026-MM-DD
+
+### Changed
+
+* [#136](https://github.com/sebastianbergmann/comparator/pull/136): Cleaner diff output for canonicalized list comparisons
+* [#160](https://github.com/sebastianbergmann/comparator/pull/160): Use by-value comparison for closures
+
+### Fixed
+
+* [#136](https://github.com/sebastianbergmann/comparator/pull/136): Canonicalized array comparison destroys string keys
+
 ## [8.2.0] - 2026-05-20
 
 ### Added
@@ -51,6 +62,17 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 ### Removed
 
 * Removed support for PHP 8.3
+
+## [7.1.8] - 2026-MM-DD
+
+### Changed
+
+* [#136](https://github.com/sebastianbergmann/comparator/pull/136): Cleaner diff output for canonicalized list comparisons
+* [#160](https://github.com/sebastianbergmann/comparator/pull/160): Use by-value comparison for closures
+
+### Fixed
+
+* [#136](https://github.com/sebastianbergmann/comparator/pull/136): Canonicalized array comparison destroys string keys
 
 ## [7.1.7] - 2026-05-20
 
@@ -372,6 +394,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 * Added `SebastianBergmann\Comparator\Factory::reset()` to unregister all non-default comparators
 * Added support for `phpunit/phpunit-mock-objects` version `^5.0`
 
+[8.2.1]: https://github.com/sebastianbergmann/comparator/compare/8.2.0...main
 [8.2.0]: https://github.com/sebastianbergmann/comparator/compare/8.1.4...8.2.0
 [8.1.4]: https://github.com/sebastianbergmann/comparator/compare/8.1.3...8.1.4
 [8.1.3]: https://github.com/sebastianbergmann/comparator/compare/8.1.2...8.1.3
@@ -379,6 +402,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 [8.1.1]: https://github.com/sebastianbergmann/comparator/compare/8.1.0...8.1.1
 [8.1.0]: https://github.com/sebastianbergmann/comparator/compare/8.0.0...8.1.0
 [8.0.0]: https://github.com/sebastianbergmann/comparator/compare/7.1...8.0.0
+[7.1.8]: https://github.com/sebastianbergmann/comparator/compare/7.1.7...7.1
 [7.1.7]: https://github.com/sebastianbergmann/comparator/compare/7.1.6...7.1.7
 [7.1.6]: https://github.com/sebastianbergmann/comparator/compare/7.1.5...7.1.6
 [7.1.5]: https://github.com/sebastianbergmann/comparator/compare/7.1.4...7.1.5